Added support for PicoJobs
