+++ gradle.properties
@@ -21,3 +21,4 @@ mcmmoVersion=b79d452a6d
 combatlogxVersion=10.0.0.0-SNAPSHOT
 factionsVersion=1.6.9.5-U0.5.23
 shopguiplusVersion=2.2.0
+picoJobsVersion=1.0-beta-2a
